+++ chapter02/linearsearch.go
@@ -1,9 +1,9 @@
-// Package chapter02 contains 
+// Package chapter02 contains
 // implementations of the algorithms introduced in Chapter 2.
 package chapter02
 
 // LinearSearch is the most basic version of linear search algorithm.
-func LinearSearch(A []int, n int, x int) int {
+func LinearSearch(A []float64, n int, x float64) int {
 
 	var answer int
