+++ chapter02/recursivelinearsearch.go
@@ -1,9 +1,9 @@
-// Package chapter02 contains 
+// Package chapter02 contains
 // implementations of the algorithms introduced in Chapter 2.
 package chapter02
 
 // RecursiveLinearSearch is a recursive style version of linear search algorithm.
-func RecursiveLinearSearch(A []int, n int, i int, x int) int {
+func RecursiveLinearSearch(A []float64, n int, i int, x float64) int {
 
 	if i > n-1 {
 		return -1
@@ -13,5 +13,5 @@ func RecursiveLinearSearch(A []int, n int, i int, x int) int {
 		return i
 	}
 
-	return RecursiveLinearSearch(A, n, i + 1, x)
+	return RecursiveLinearSearch(A, n, i+1, x)
 }
